Standalone inference complete
