+++ configs/eval_cm4_thermo.yaml
@@ -2,8 +2,9 @@ debug: true
 disk_mode: true
 inference:
   start_time: 311-01-01
-  end_time: 312-01-01 # 351-01-01
-ckpt_path: /pscratch/sd/s/suryad/Ocean_Emulator/.LOCAL/2025-02-09-train-local_cm4_samudra_thermo/saved_nets/convnextunet_epoch_3_best_validation_steps_4_all_Lateral_Data_025_no_smooth.pt
+  end_time: 351-01-01
+ckpt_path: /pscratch/sd/s/suryad/Ocean_Emulator/temp/convnextunet_epoch_36_best_inference_steps_4_all_Lateral_Data_025_no_smooth.pt
+num_model_steps_forward: 20
 record_every: 10
 
 experiment:
@@ -12,15 +13,16 @@ experiment:
   rand_seed: 15
   base_output_dir: .LOCAL
   gantry: false
-  cluster_data_dir: /pscratch/sd/s/suryad/data # temporary for notebook testing
+  cluster_data_dir: null
   wandb:
     mode: online
-    project: 3D_ocean_emu_CM4
-    entity: suryadheeshjith
+    project: ocean-emulator
+    entity: ai2cm
+    group: samudra-inference-2in2out
   network: convnextunet
-  exp_num_in: 3D_noFast_all
+  exp_num_in: 3D_all
   exp_num_extra: 3D_all
-  exp_num_out: 3D_noFast_all
+  exp_num_out: 3D_all
 
 data:
   wet_file: CM4_5daily_v0.4.0_wetmask
