+++ scripts/.local-eval
@@ -12,4 +12,4 @@ echo 'Using config file:' $CONFIG_PATH
 TMP_CONFIG_PATH=/tmp/config_temp.yaml
 sed "s|cluster_data_dir:.*|cluster_data_dir: $OCEAN_DATA_DIR|" $CONFIG_PATH > $TMP_CONFIG_PATH
 
-python3 src/eval.py --config $TMP_CONFIG_PATH
+python3 src/eval.py --config $TMP_CONFIG_PATH --subname "thermodynamic-FusedAdam-lr0.0001-scheduler-epoch36/120-infchmean0.073313"
