+++ src/config.py
@@ -201,6 +201,7 @@ class EvalConfig:
     debug: bool = False
     disk_mode: bool = True
     ckpt_path: str = ""
+    num_model_steps_forward: int = 200
     record_every: int = 10
     # Config components
     inference: TimeConfig = field(
@@ -234,6 +235,7 @@ class EvalConfig:
             "debug": self.debug,
             "disk_mode": self.disk_mode,
             "ckpt_path": self.ckpt_path,
+            "num_model_steps_forward": self.num_model_steps_forward,
             "record_every": self.record_every,
             "inference": self.inference.__dict__,
             "experiment": self.experiment.__dict__,
